complete backend (finished all api)

+++ backend/models/recipe.function.js
@@ -32,13 +32,19 @@ export const addNewRecipe = async (req, res) => {
 export const deleteRecipe = async (req, res) => {
     const {id} = req.params;
     console.log("ID to be deleted: ", id);
+
+    if (mongoose.Types.ObjectId.isValid(id) === false) {
+        res.status(404).json({success: false, message: "Recipe not found"});
+        return;
+    }
+
     try {
         await Recipe.findByIdAndDelete(id).then(doc => console.log("Deleted recipe: ", doc));
         res.status(200).json({success: true, message: "Recipe deleted successfully"});
         await Recipe.find().then(recipes => console.log("Remaining recipes: ", recipes));
     } catch (error) {
         console.error("Error in delete recipe: ", error.message);
-        res.status(404).json({success: false, message: "Recipe not found"});
+        res.status(500).json({success: false, message: "Server error"});
     }
 }
 
@@ -55,6 +61,22 @@ export const updateRecipe = async (req, res) => {
         await Recipe.findByIdAndUpdate(id, recipe, {new: true}).then(doc => console.log("Updated recipe: ", doc));  
         res.status(200).json({success: true, message: "Recipe updated successfully"});
     } catch (error) {
+        res.status(500).json({success: false, message: "Server error"});
+    }
+}
+
+export const getRecipe = async (req, res) => {
+    const {id} = req.params;
+    if (mongoose.Types.ObjectId.isValid(id) === false) {
+        res.status(404).json({success: false, message: "Recipe not found"});
+        return;
+    }
+
+    try {
+        const recipe = await Recipe.findById(id);
+        res.status(200).json({success: true, data: recipe});
+    } catch (error) {
+        console.error("Error in get recipe: ", error.message);
         res.status(500).json({success: false, message: "Server error"});
     }
 }
